Make parsing bool values in filter optional
This ensures that string values containing boolean values in a filter are
not parsed as bool by default. Otherwise the filter might not be
correct since we do not know if the string is intended or not.
The intended parsing of certain fields can be forced by setting a list
of fields ('boolFilterFields'), which will be tried to parse as boolean.